refactor: remove custom `Match` class
Now that I know there's a JSON representation as part of RipGrep, it doesn't make sense to maintain an in-house `Match` class at all; we should just use the exact output from RipGrep instead.

BREAKING CHANGE: The output of the search command is now different; it's a JSON representation that matches the format that RipGrep itself produces exactly.